+++ CHANGES.md
@@ -1,5 +1,12 @@
 Change Log
 ==========
 
-## changes in 1.1-SNAPSHOT
+## changes in 1.2-SNAPSHOT
 
+
+## changes in 1.1
+
+* requires scala.js-0.6.14
+* removed dependency on jsext, see: https://github.com/jducoeur/jsext
+* changed DatepickerOptions to use scala.js-0.6.14 configuration objects
+